Fix handling unified NPG errors
